ibook/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
pbr==5.6.0
platformdirs==2.3.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.2
stevedore==3.4.0
toml==0.10.2
uritemplate==3.0.1
virtualenv==20.8.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
